Add deletion and checking to the RBtree
Add deletion to the RBtree via a method erase(). To realize it was necessary
to change behaviour of the bindTo* methods of the Node. Since this
version it automaticaly delete the previous binded node, instead of
to throw exception.
Also implement some tests: add to RBtree a method check() that checks
propriety of the red-black conditions.