Improve profiler, get first pretty graphs
Make measure time using one tree and insert a large number of elems (near 10000) to it.
This way get just average time statistics, beacuse this quantity is large, but very less than
amount of the tree elements.
Also make graph builder in the graph.py based on pandas and matplotlib.
Also detect and fix a bug in the RBtree: number of elements not decrease in
every cases of deletion.